requirements.txt
python-telegram-bot[job-queue]==21.6
aiohttp==3.10.11
beautifulsoup4==4.12.3
lxml==5.3.0
python-dotenv==1.0.1
pydantic==2.10.3
pydantic-settings==2.7.0
cachetools==5.5.0
pytz==2024.2
